+++ components/utilities/header.js
@@ -9,12 +9,12 @@ import Image from "next/image";
 import Menu from "../utilities/menu";
 export default function Header() {
   return (
-    <div className="py-5 mb-2 w-screen   bg-[url('/assets/menu/menu.webp')]">
-      <Head>
+    <div className="bg-[#f8fafc] h-10 bg-opacity-30 flex">
+      <span>
         <title>SSN IEEESB</title>
         <meta name="description" content="SSN IEEE Student Branch" />
         <link rel="icon" href="/favicon.ico" />
-      </Head>
+      </span>
 
       <Menu></Menu>
     </div>
