+++ components/utilities/menu-item.js
@@ -9,10 +9,10 @@ import Image from "next/image";
 
 export default function MenuItem({ url, name }) {
   return (
-    <div className="inline p-5 text-[150%] m-10">
+    <span className="inline p-5 text-[150%] m-10">
       <a className="hover:bg-[#ddd] hover:text-[black]" href={url}>
         {name}
       </a>
-    </div>
+    </span>
   );
 }
