+++ tailwind.config.js
@@ -8,6 +8,7 @@
 // }
 /** @type {import('tailwindcss').Config} */
 module.exports = {
+  darkMode: ["class", '[data-mode="dark"]'],
   content: [
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
